add changes - move all errors to single file - update error handling in worker handler - remove logfile from git
