+++ .gitignore
@@ -1,2 +1,3 @@
 .env
 /.vscode
+logfile
