--- logfile
@@ -1,75 +1,0 @@
-{"L":"INFO","T":"2025-02-05T17:36:13.180+0530","C":"cmd/main.go:50","M":"Custom logger initialized"}
-2025-02-05T17:36:31.014+0530	INFO	cmd/main.go:50	Custom logger initialized
-2025-02-05T17:39:52.352+0530	ERROR	logger/logger.go:35	Server is Up and Running	{"port": "8080"}
-github.com/harsh-jagtap-josh/RozgarLink/internal/pkg/logger.Errorw
-	/media/harsh/New Volume/JOSH/RozgarLink/internal/pkg/logger/logger.go:35
-main.main
-	/media/harsh/New Volume/JOSH/RozgarLink/cmd/main.go:42
-runtime.main
-	/usr/local/go/src/runtime/proc.go:272
-2025-02-05T17:40:13.277+0530	ERROR	logger/logger.go:35	Server is Up and Running	{"port": "8080"}
-github.com/harsh-jagtap-josh/RozgarLink/internal/pkg/logger.Errorw
-	/media/harsh/New Volume/JOSH/RozgarLink/internal/pkg/logger/logger.go:35
-main.main
-	/media/harsh/New Volume/JOSH/RozgarLink/cmd/main.go:42
-runtime.main
-	/usr/local/go/src/runtime/proc.go:272
-{"port": "8080"}
-2025-02-05T17:42:30.870+0530	ERROR	logger/logger.go:35	Server is Up and Running	{"port": "8080"}
-github.com/harsh-jagtap-josh/RozgarLink/internal/pkg/logger.Errorw
-	/media/harsh/New Volume/JOSH/RozgarLink/internal/pkg/logger/logger.go:35
-main.main
-	/media/harsh/New Volume/JOSH/RozgarLink/cmd/main.go:42
-runtime.main
-	/usr/local/go/src/runtime/proc.go:272
-2025-02-05T17:43:54.959+0530	INFO	logger/logger.go:39	Server is Up and Running	{"port": "8080"}
-2025-02-05T17:44:20.207+0530	INFO	logger/logger.go:39	Server is Up and Running	{"port": "8080"}
-2025-02-05T17:44:32.048+0530	ERROR	logger/logger.go:35	Server is Up and Running	{"port": "8080"}
-github.com/harsh-jagtap-josh/RozgarLink/internal/pkg/logger.Errorw
-	/media/harsh/New Volume/JOSH/RozgarLink/internal/pkg/logger/logger.go:35
-main.main
-	/media/harsh/New Volume/JOSH/RozgarLink/cmd/main.go:42
-runtime.main
-	/usr/local/go/src/runtime/proc.go:272
-2025-02-05T17:44:44.111+0530	ERROR	logger/logger.go:35	Server is Up and Running	{"port": "8080"}
-github.com/harsh-jagtap-josh/RozgarLink/internal/pkg/logger.Errorw
-	/media/harsh/New Volume/JOSH/RozgarLink/internal/pkg/logger/logger.go:35
-main.main
-	/media/harsh/New Volume/JOSH/RozgarLink/cmd/main.go:42
-runtime.main
-	/usr/local/go/src/runtime/proc.go:272
-2025-02-05T17:45:02.209+0530	ERROR	logger/logger.go:35	Server is Up and Running	{"port": "8080"}
-github.com/harsh-jagtap-josh/RozgarLink/internal/pkg/logger.Errorw
-	/media/harsh/New Volume/JOSH/RozgarLink/internal/pkg/logger/logger.go:35
-main.main
-	/media/harsh/New Volume/JOSH/RozgarLink/cmd/main.go:42
-runtime.main
-	/usr/local/go/src/runtime/proc.go:272
-2025-02-05T17:45:19.071+0530	ERROR	logger/logger.go:35	Server is Up and Running	{"port": "8080"}
-github.com/harsh-jagtap-josh/RozgarLink/internal/pkg/logger.Errorw
-	/media/harsh/New Volume/JOSH/RozgarLink/internal/pkg/logger/logger.go:35
-main.main
-	/media/harsh/New Volume/JOSH/RozgarLink/cmd/main.go:42
-runtime.main
-	/usr/local/go/src/runtime/proc.go:272
-1.738757756445198e+09	error	logger/logger.go:35	Server is Up and Running	{"port": "8080"}
-github.com/harsh-jagtap-josh/RozgarLink/internal/pkg/logger.Errorw
-	/media/harsh/New Volume/JOSH/RozgarLink/internal/pkg/logger/logger.go:35
-main.main
-	/media/harsh/New Volume/JOSH/RozgarLink/cmd/main.go:42
-runtime.main
-	/usr/local/go/src/runtime/proc.go:272
-2025-02-05T17:47:42.917+0530	ERROR	logger/logger.go:35	Server is Up and Running	{"port": "8080"}
-github.com/harsh-jagtap-josh/RozgarLink/internal/pkg/logger.Errorw
-	/media/harsh/New Volume/JOSH/RozgarLink/internal/pkg/logger/logger.go:35
-main.main
-	/media/harsh/New Volume/JOSH/RozgarLink/cmd/main.go:42
-runtime.main
-	/usr/local/go/src/runtime/proc.go:272
-2025-02-05T17:48:51.301+0530	ERROR	logger/logger.go:35	Server is Up and Running	{"port": "8080"}
-github.com/harsh-jagtap-josh/RozgarLink/internal/pkg/logger.Errorw
-	/media/harsh/New Volume/JOSH/RozgarLink/internal/pkg/logger/logger.go:35
-main.main
-	/media/harsh/New Volume/JOSH/RozgarLink/cmd/main.go:42
-runtime.main
-	/usr/local/go/src/runtime/proc.go:272
